flex and style
