+++ client/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "files.exclude": {
+    "**/**/**/**/**/**/**/**.test.tsx": true,
+    "**/**/**/**/**/**/**/**.css.map": true
+  }
+}
