+++ client/src/components/app/App.css
@@ -1,0 +1,12 @@
+.App {
+  text-align: center;
+  font-family: Droid Sans, monospace;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+}
+/*# sourceMappingURL=App.css.map */
