+++ client/src/components/mainview/Mainview.css
@@ -1,0 +1,24 @@
+.Mainview {
+  width: 100vw;
+  height: 100vh;
+  background-color: #222226;
+  color: wheat;
+  font-size: 1.7vh;
+  -ms-flex-line-pack: center;
+      align-content: center;
+}
+
+.InputForms {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+  -webkit-box-pack: space-evenly;
+      -ms-flex-pack: space-evenly;
+          justify-content: space-evenly;
+  margin-top: 10vh;
+}
+/*# sourceMappingURL=Mainview.css.map */
