+++ client/src/components/mainview/forms/FormJS.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import "./FormJS.css"
+
+const FormJS: React.FC = () => {
+  return (
+    <div className="FormJS">
+        <textarea className="InputForm">
+        </textarea>
+    </div>
+  );
+}
+
+export default FormJS;
