+++ client/src/components/mainview/forms/FormTS.css
@@ -1,0 +1,7 @@
+.FormTS {
+  background-color: #333336;
+  width: 40vw;
+  height: 70vh;
+  border-radius: 1vh;
+}
+/*# sourceMappingURL=FormTS.css.map */
