+++ client/src/components/mainview/forms/FormTS.scss
@@ -1,0 +1,8 @@
+$theme-color: #333336;
+
+.FormTS {
+  background-color: $theme-color;
+  width: 40vw;
+  height: 70vh;
+  border-radius: 1vh;
+}
