--- client/src/components/mainview/inputforms/FormJS.scss
@@ -1,5 +1,0 @@
-.FormJS {
-  background-color: black;
-  width: 40vw;
-  height: 70vh;
-}
