--- client/src/components/mainview/inputforms/FormTS.scss
@@ -1,5 +1,0 @@
-.FormTS {
-  background-color: black;
-  width: 40vw;
-  height: 70vh;
-}
