--- client/src/components/mainview/inputforms/FormTS.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import "./FormTS.scss"
-
-const FormTS: React.FC = () => {
-  return (
-    <div className="FormTS">
-    </div>
-  );
-}
-
-export default FormTS;
