+++ client/src/components/navbar/Navbar.css
@@ -1,0 +1,50 @@
+.Navbar {
+  width: 75px;
+  height: 100vh;
+  background-color: #333336;
+  color: wheat;
+  -webkit-transform-origin: 50% 65%;
+          transform-origin: 50% 65%;
+  -webkit-transition: -webkit-transform 0.7s, -webkit-filter 0.5s ease-in-out;
+  transition: -webkit-transform 0.7s, -webkit-filter 0.5s ease-in-out;
+  transition: transform 0.7s, filter 0.5s ease-in-out;
+  transition: transform 0.7s, filter 0.5s ease-in-out, -webkit-transform 0.7s, -webkit-filter 0.5s ease-in-out;
+  -webkit-filter: brightness(100%);
+          filter: brightness(100%);
+}
+
+.Navbar img {
+  margin-top: 0.5vh;
+  margin-bottom: 0.5vh;
+  margin-left: 0.5vw;
+  margin-right: 0.5vw;
+  max-width: 50px;
+  max-height: 50px;
+}
+
+.Navbar:hover {
+  -webkit-filter: brightness(130%);
+          filter: brightness(130%);
+  -webkit-transform: scale(1.3);
+          transform: scale(1.3);
+}
+
+.Logos {
+  margin-top: 15vh;
+  height: 40vh;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -ms-flex-pack: distribute;
+      justify-content: space-around;
+}
+
+.ImageLink:hover img {
+  -webkit-box-shadow: 0 0 5px #7289DA;
+          box-shadow: 0 0 5px #7289DA;
+}
+/*# sourceMappingURL=Navbar.css.map */
