+++ client/src/components/navbar/Navbar.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Navbar.scss"
+import "./Navbar.css"
 import vklogo from "../../resourses/logos/VK-logo.png"
 import settingslogo from "../../resourses/logos/settingslogo.png"
 
@@ -7,8 +7,12 @@ const Navbar: React.FC = () => {
   return (
     <div className="Navbar">
       <div className="Logos">
-        <img src={vklogo} alt="Logo"/>
-        <img src={settingslogo} alt="Logo"/>
+        <a className="ImageLink" href="https://vk.com/keksecproduct">
+          <img src={vklogo} alt="Logo"/>
+        </a>
+        <a className="ImageLink" href="https://vk.com/keksecproduct">
+          <img src={settingslogo} alt="Logo"/>
+        </a>
       </div>
     </div>
   );
